requirements.txt
loguru==0.7.3
openai==1.58.1
python-dotenv==1.0.1
tenacity==9.0.0
rouge-chinese
jieba==0.42.1
numpy<2
faiss-gpu
nltk
torch
transformers
FlagEmbedding